Move function; import at top of file rather than inside function
